Fix : 코드 가독성 향상 및 Manager -> Controller 네이밍 변경
